correct paths

+++ images/03.ServiceImages/smtp/home/engines/scripts/actionators/sudo/_import_domain_dkim.sh
@@ -3,8 +3,16 @@
  
 domain_name=$1
 
+
 required_values="domain_name"
 check_required_values
+
+if test -f /etc/opendkim/keys/$domain_name/mail.private
+ then
+  echo '{"error":"Domain key exists for '$domain_name'"}'
+  exit 2
+fi
+
 domain_dir=/etc/opendkim/keys/$domain_name
 mkdir -p $domain_dir
 cat - > $domain_dir/mail.private
@@ -13,3 +21,5 @@ chown opendkim -R $domain_dir
 mv /tmp/public > $domain_dir/mail.txt
 chmod g+r $domain_dir/mail.txt
 
+ echo '{"Success":"Domain key for '$domain_name' Imported"}'
+exit 0
